feat: authenticate to Metamask wallet
Use Web3 provider to operate with the same connection
instance across the app.

Add a fork between a connected and not connected wallet
cases.

Add a button component and the respective utils
to connect to Metamask wallet on user click

The connection is picked up automatically on initial
render of the app if exists