feat: add selected chain validator and listeners
In case if an unsupported chain is selected
the application will show a different UI for the case.

A wallet context has been created
to set the initial wallet values,
and also listen to account change and
chain switch events. This is required
to show a separate UI for the case.

Also a validator for the connected wallet was added.
Currently, only Metamask is supported