feat2: fetch and render the crypto assets
The assets are fetched from the connected wallet.

The native token of the selected chain is always fetched and rendered.

The other tokens are fetched by token addresses from the array constant.
This array contains tokens from differnt chains.
We check if the token belongs to the selected chain by calling
the totalSupply method from it's ABI. If the call of totalSupply fails
it is considered that the token doesn't belong to the current chain.
These tokens are filtered from data for rendering the assets.
A warning in console is fired for the failed tokens.

CHR token is included in the array of token addresses.